% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hail.R
\docType{package}
\name{hail}
\alias{hail}
\alias{hail-package}
\title{Retrieve Data from the HYDRA Project}
\description{
Retrieves rainfall data from the City of Portland's HYDRA monitoring stations.
}
\seealso{
\code{\link{hail_hydra}} for data retrieval.
}

