% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hail.R
\docType{data}
\name{hydra_data}
\alias{hydra_data}
\title{HYDRA metadata}
\format{A data frame with 53 rows and 3 variables:
\describe{
  \item{station}{The station name. This can be passed into \code{\link{hail_hydra}} to
  retrieve data from that station.}
  \item{address}{The station's address/}
  \item{url}{The URL of the station's raw, unread data/}
}}
\source{
\url{http://or.water.usgs.gov/precip/}
}
\usage{
hydra_data
}
\description{
metadata about the HYDRA rain gages.
}
\keyword{datasets}

