% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ hail }
\alias{ hail }
\alias{ hail }
\title{ Retrieve Data from the HYDRA Project }
\description{
Retrieves rainfall data from the City of Portland's &quot;HYDRA&quot; monitoring stations.
}
